requirements.txt
flask==3.0.2
google-generativeai==0.3.2 